fix second useCase SaveAndQuoteText - before incorporate TextQuoted in TextQuoteMade

+++ src/domain/model/src/main/java/com/libraryproviderbackend/user/UserBehavior.java
@@ -9,19 +9,23 @@ import com.libraryproviderbackend.user.values.user.Email;
 import com.libraryproviderbackend.user.values.user.EntryDate;
 import com.libraryproviderbackend.user.values.user.Password;
 
+import java.math.BigDecimal;
+
 public class UserBehavior extends EventChange {
 
     DiscountsEnum[] discountsEnum = DiscountsEnum.values();
     TextTypeEnum[] textTypeEnum = TextTypeEnum.values();
 
     public UserBehavior(User user){
+
         addSubscriber(UserCreated.class, event -> {
             user.email = Email.of(event.getEmail());
             user.password = Password.of(event.getPassword());
             user.entryDate = EntryDate.of(String.valueOf(event.getEntryDate()));
         });
+
         addSubscriber(TextSavedAndQuoted.class, event -> {
-            user.quoteText(event.title, event.initialPrice, textTypeEnum[event.type], DiscountsEnum.NONE);
+            user.quoteText(event.title, BigDecimal.valueOf(event.initialPrice), textTypeEnum[event.type], DiscountsEnum.NONE);
         });
     }
 }
